style: WebSocket Package 위치 이동
